Add support for creating a hidden window
This helps improve the load experience for complex UIs that need time to
finish rendering. Recommend calling an RPC from JavaScript to tell the
Go code when it's ready to be shown.

On Mac, we assume we are running as an LSUIElement app, as we change the
activation policy to show/hide the Dock icon corresponding to the
window's visibility.

GTK untested..